backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pandas
kiteconnect
ollama
python-dotenv
aiohttp
